Mock test class
