+++ agent/src/test/java/com/microsoft/hydralab/agent/runner/ActionExecutorTest.java
@@ -2,69 +2,46 @@
 // Licensed under the MIT License.
 package com.microsoft.hydralab.agent.runner;
 
-import com.alibaba.fastjson.JSONArray;
 import com.alibaba.fastjson.JSONObject;
 import com.microsoft.hydralab.agent.test.BaseTest;
 import com.microsoft.hydralab.common.entity.common.DeviceAction;
 import com.microsoft.hydralab.common.entity.common.DeviceInfo;
-import com.microsoft.hydralab.common.entity.common.Message;
-import com.microsoft.hydralab.common.management.DeviceManager;
-import com.microsoft.hydralab.common.util.SerializeUtil;
-import org.junit.jupiter.api.Assertions;
+import com.microsoft.hydralab.common.management.impl.AndroidDeviceManager;
 import org.junit.jupiter.api.Test;
+import org.mockito.Mockito;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 
 class ActionExecutorTest extends BaseTest {
 
     ActionExecutor actionExecutor = new ActionExecutor();
 
     @Test
-    void doAction() {
-        DeviceManager deviceManager = new MockDeviceManager();
+    void createAndExecuteActions() {
+        AndroidDeviceManager deviceManager = Mockito.mock(AndroidDeviceManager.class);
         DeviceInfo deviceInfo = new DeviceInfo();
         JSONObject actionJson = new JSONObject();
         actionJson.put("method", "setProperty");
         actionJson.put("deviceType", "Android");
-
         DeviceAction action1 = JSONObject.parseObject(actionJson.toJSONString(), DeviceAction.class);
-        action1.getArgs().add("paramA");
-        action1.getArgs().add("paramB");
+        List<String> args1 = List.of("paramA", "paramB");
+        action1.setArgs(args1);
         actionExecutor.doAction(deviceManager, deviceInfo, baseLogger, action1);
+        verify(deviceManager).setProperty(deviceInfo, args1.get(0), args1.get(1), baseLogger);
 
         DeviceAction action2 = new DeviceAction("Android", "changeGlobalSetting");
-        action2.getArgs().add("paramC");
-        action2.getArgs().add("paramD");
+        List<String> args2 = List.of("paramC", "paramD");
+        action2.setArgs(args2);
         List<DeviceAction> actions = new ArrayList<>();
         actions.add(action1);
         actions.add(action2);
         actionExecutor.doActions(deviceManager, deviceInfo, baseLogger, Map.of(DeviceAction.When.SET_UP, actions), DeviceAction.When.SET_UP);
-
-    }
-
-    @Test
-    void testJsonArraySerialize() {
-        JSONArray array = new JSONArray();
-        array.add(1);
-        array.add(3);
-        array.add(5);
-        array.add(2);
-        array.add(4);
-        array.add(6);
-        String str1 = array.toJSONString();
-        baseLogger.info(str1);
-        Message msg = new Message();
-        msg.setBody(array);
-        Message msg2 = SerializeUtil.byteArrToMessage(SerializeUtil.messageToByteArr(msg));
-        String str2 = ((JSONArray) msg2.getBody()).toJSONString();
-        baseLogger.info(str2);
-        JSONArray array2 = JSONArray.parseArray(str2);
-        String str3 = array2.toJSONString();
-        baseLogger.info(str3);
-        Assertions.assertEquals(str1, str2, "Serialize error!");
-        Assertions.assertEquals(str1, str3, "Serialize error!");
+        verify(deviceManager, times(2)).setProperty(deviceInfo, args1.get(0), args1.get(1), baseLogger);
 
     }
